+++ packages/blockprovider-example/server.js
@@ -8,7 +8,7 @@ const app = express();
 const config = require('common-env/withLogger')(console).getOrElseAll({
   port: 3000,
 
-  // from where we can access this server from outside (e.g. "https://mydomain.com:8083") we access to access ti
+  // where this server can be reached from outside (e.g. "https://mydomain.com:8083")
   endpoint: 'http://localhost:3000',
 });
 
@@ -33,7 +33,7 @@ app.get('/configurations', (req, res) =>
             // we could also specify a FQDN
             url: `/block/bitcoin`,
             method: 'GET',
-            // we block depends on time so it's not pure and won't be cached for a long time
+            // this block depends on time, so it's not pure and won't be cached for a long time
             pure: false,
             required: ['threshold'],
             parameters: [