+++ packages/blockprovider-tutorial-parkings/src/main/java/fr/ouestfrance/platform/domain/ParkData.java
@@ -1,0 +1,52 @@
+package fr.ouestfrance.platform.domain;
+
+public class ParkData {
+
+    private String name;
+    private boolean full = false;
+    private boolean closed = false;
+    private int available;
+    private int max;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public boolean isFull() {
+        return full;
+    }
+
+    public void setFull(boolean full) {
+        this.full = full;
+    }
+
+    public boolean isClosed() {
+        return closed;
+    }
+
+    public void setClosed(boolean closed) {
+        this.closed = closed;
+    }
+
+    public int getAvailable() {
+        return available;
+    }
+
+    public void setAvailable(int available) {
+        this.available = available;
+    }
+
+    public int getMax() {
+        return max;
+    }
+
+    public void setMax(int max) {
+        this.max = max;
+    }
+}
+
+